Untextured raycasting working, but coordinates mirrored

+++ srcs/player.h
@@ -6,7 +6,7 @@
 /*   By: lchapren <lchapren@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/19 13:36:00 by lchapren          #+#    #+#             */
-/*   Updated: 2020/07/21 13:39:44 by lchapren         ###   ########.fr       */
+/*   Updated: 2020/07/22 09:43:03 by lchapren         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ typedef struct	s_player
 	float	position_y;
 	float	direction_x;
 	float	direction_y;
-	int		plane_x;
-	int		plane_y;
+	float	plane_x;
+	float	plane_y;
 	
 	float	speed;
 	float	rotation_angle;
